pyproject.toml
[project]
name = "arac"
version = "0.1.0"
description = "AkashicRecords Agent Client - Multi-agent system for file-based knowledge management"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-adk",
    "litellm>=1.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.0.0",
    "mypy>=1.0.0"
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.0.0",
    "mypy>=1.0.0"
]
